@@ -11,27 +11,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.14",
-    "@angular/cdk": "^10.1.3",
-    "@angular/common": "^10.0.14",
-    "@angular/compiler": "^10.0.14",
-    "@angular/compiler-cli": "^10.0.14",
-    "@angular/core": "^10.0.14",
-    "@angular/forms": "^10.0.14",
-    "@angular/localize": "^10.0.14",
-    "@angular/material": "^10.1.3",
-    "@angular/platform-browser": "^10.0.14",
-    "@angular/platform-browser-dynamic": "^10.0.14",
-    "@angular/router": "^10.0.14",
-    "@ng-bootstrap/ng-bootstrap": "^7.0.0",
-    "bootstrap": "^4.5.2",
-    "jquery": "^3.5.1",
-    "ngb": "^1.1.0",
-    "popper.js": "^1.16.1",
-    "rxjs": "^6.6.2",
-    "tslib": "^2.0.1",
-    "typescript": "^4.0.2",
-    "zone.js": "^0.11.1"
+    "@angular/animations": "10.1.1",
+    "@angular/cdk": "10.2.1",
+    "@angular/common": "10.1.1",
+    "@angular/compiler": "10.1.1",
+    "@angular/compiler-cli": "10.1.1",
+    "@angular/core": "10.1.1",
+    "@angular/forms": "10.1.1",
+    "@angular/localize": "10.1.1",
+    "@angular/material": "10.2.1",
+    "@angular/platform-browser": "10.1.1",
+    "@angular/platform-browser-dynamic": "10.1.1",
+    "@angular/router": "10.1.1",
+    "@ng-bootstrap/ng-bootstrap": "7.0.0",
+    "@ngrx/effects": "10.0.0",
+    "@ngrx/store": "10.0.0",
+    "bootstrap": "4.5.2",
+    "jquery": "3.5.1",
+    "ngb": "1.1.0",
+    "ngrx-store-freeze": "^0.2.4",
+    "ngrx-store-logger": "^0.2.4",
+    "popper.js": "1.16.1",
+    "rxjs": "6.6.3",
+    "tslib": "2.0.1",
+    "typescript": "4.0.2",
+    "zone.js": "0.11.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.1000.5",
@@ -51,6 +55,9 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.9.5"
+    "typescript": "~3.9.5",
+    "@ngrx/store-devtools": "^10.0.0",
+    "ngrx-store-logger": "^0.2.4",
+    "ngrx-store-freeze": "^0.2.4"
   }
 }
